# backend/Services/swagger.yaml
openapi: 3.0.3
info:
  title: fashion_store API
  version: 1.0.0
  description: API สำหรับจัดการสินค้า ทั้งส่วน Admin และ User
servers:
  - url: http://localhost:3000
...
paths:
...
  /products/{productId}:
    delete:
      tags:
        - Products
      summary: ลบสินค้า (Admin)
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: ลบสินค้าสำเร็จ
          content:
            application/json:
              example: { message: "DELETED" }
        '404':
          description: ไม่พบสินค้า
          content:
            application/json:
              example: { message: "Product not Found" }
        '500':
          description: Server error
          content:
            application/json:
              example: { message: "Server error deleteProduct" }

  # =========================
  # User / General Endpoints
  # =========================
...
  /products/{id}:
    get:
      tags:
        - Products
      summary: ดูรายละเอียดสินค้า 1 ชิ้น
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: รายละเอียดสินค้า
          content:
            application/json:
              example:
                product_id: 1
                name: "รองเท้า Nike"
                description: "รองเท้าวิ่ง"
                price: 2500
                stock_quantity: 10
                image_filename: "nike_air"
                brand: "Nike"
                category_name: "Running"
                detail: "รองเท้าวิ่งรุ่นใหม่ล่าสุด"
        '500':
          description: Server error
          content:
            application/json:
              example: { message: "Server error getProductId" }

  # =========================
  # Special Endpoints
  # =========================
...